+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## v3.0.0 - 2024-10-08
 ### What's Changed
-**Full Changelog**: https://github.com/obervinov/pyinstabot-downloader/compare/v2.3.0...v3.0.0-beta by @obervinov in https://github.com/obervinov/pyinstabot-downloader/pull/101
+**Full Changelog**: https://github.com/obervinov/pyinstabot-downloader/compare/v2.3.0...v3.0.0 by @obervinov in https://github.com/obervinov/pyinstabot-downloader/pull/101
 #### 💥 Breaking Changes
 * replacement of [instaloder](https://github.com/instaloader/instaloader) module with [instagrapi](https://github.com/subzeroid/instagrapi) module
 * module `proxying` configuration moved to Vault