+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * change of parameter configuration for `downloader-api` and `uploader-api` functions. Please, check the new parameters in [README.md](README.md#bot-configuration-source-and-supported-parameters)
 #### 🚀 Features
 * replacement of [instaloder](https://github.com/instaloader/instaloader) module with [instagrapi](https://github.com/subzeroid/instagrapi) module
+* replace the base image of project with `python:3.9.20` (Ubuntu instead of Alpine)
 
 
 ## v2.3.0 - 2024-10-04