+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 **Full Changelog**: https://github.com/obervinov/pyinstabot-downloader/compare/v2.3.0...v3.0.0 by @obervinov in https://github.com/obervinov/pyinstabot-downloader/pull/102
 #### 💥 Breaking Changes
 * replacement of [instaloder](https://github.com/instaloader/instaloader) module with [instagrapi](https://github.com/subzeroid/instagrapi) module
-* module `proxying` configuration moved to Vault
+* function with the `proxy' configuration has been moved to Vault
 * remove the outdated storage types: `Dropbox` and `Mega` (Support for individual cloud providers removed in favour of `webdav` compatible providers)
 * change of parameter configuration for `downloader-api` and `uploader-api` in the Vault. Please, check the new parameters in [README.md](README.md#bot-configuration-source-and-supported-parameters)
 #### 🚀 Features